Fix #14 and #16 : Seperated SPI and NRF functionalities
