+++ PSLab_Original/nbproject/configurations.xml
@@ -9,6 +9,8 @@
       <itemPath>PSLAB_UART.h</itemPath>
       <itemPath>Common_Methods.h</itemPath>
       <itemPath>PSLAB_I2C.h</itemPath>
+      <itemPath>PSLAB_NRF.h</itemPath>
+      <itemPath>PSLAB_SPI.h</itemPath>
     </logicalFolder>
     <logicalFolder name="LinkerScript"
                    displayName="Linker Files"
@@ -21,6 +23,8 @@
       <itemPath>PSLAB_UART.c</itemPath>
       <itemPath>Common_Methods.c</itemPath>
       <itemPath>PSLAB_I2C.c</itemPath>
+      <itemPath>PSLAB_NRF.c</itemPath>
+      <itemPath>PSLAB_SPI.c</itemPath>
     </logicalFolder>
     <logicalFolder name="ExternalFiles"
                    displayName="Important Files"
