+++ PSLab_Original/nbproject/private/private.xml
@@ -3,14 +3,6 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/1"/>
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/Common_Methods.h</file>
-            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/PSLAB_I2C.c</file>
-            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/PSLAB_I2C.h</file>
-            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/functions.c</file>
-            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/functions.h</file>
-            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/proto2_main.c</file>
-            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/Common_Methods.c</file>
-        </group>
+        <group/>
     </open-files>
 </project-private>
